safari-600.7-branch/LayoutTests/platform/ios-sim/media/no-autoplay-with-user-gesture-requirement-expected.txt

Test that media autoplay should not work if user gesture is required for playback.

EXPECTED (video.paused == 'true') OK
EVENT(canplaythrough)
EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
END OF TEST
